slug, image and user spec added

+++ quotes/views.py
@@ -3,6 +3,7 @@ from django.http import Http404
 from .models import Quote
 from .forms import QuoteForm
 from django.contrib import messages
+from django.utils.text import slugify
 
 # Create your views here.
 
@@ -13,8 +14,8 @@ def index(request):
     }
     return render(request, 'quotes/list.html', context)
 
-def detail(request, quote_id):
-    quote = get_object_or_404(Quote, pk = quote_id)
+def detail(request, slug):
+    quote = get_object_or_404(Quote, slug = slug)
     context = {
         'quote' : quote
     }
@@ -24,7 +25,10 @@ def search(request):
     return render(request, 'quotes/search.html')
 
 def create(request):
-    form = QuoteForm(request.POST or None)
+    if not request.user.is_authenticated:
+        raise Http404
+
+    form = QuoteForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         quote = form.save()
         messages.success(request, 'Başarılı bir şekilde oluşturdunuz')
@@ -34,9 +38,11 @@ def create(request):
     }
     return render(request, 'quotes/form.html', context)
 
-def update(request, quote_id):
-    quote = get_object_or_404(Quote, pk = quote_id)
-    form = QuoteForm(request.POST or None, instance=quote)
+def update(request, slug):
+    if not request.user.is_authenticated:
+        raise Http404
+    quote = get_object_or_404(Quote, slug = slug)
+    form = QuoteForm(request.POST or None, request.FILES or None, instance=quote)
     if form.is_valid():
         form.save()
         messages.success(request, 'Başarılı bir şekilde oluşturdunuz')
@@ -47,7 +53,9 @@ def update(request, quote_id):
     return render(request, 'quotes/form.html', context)
 
 
-def delete(request, quote_id):
-    quote = get_object_or_404(Quote, pk = quote_id)
+def delete(request, slug):
+    if not request.user.is_authenticated:
+        raise Http404
+    quote = get_object_or_404(Quote, slug = slug)
     quote.delete()
     return redirect('/quotes')
